make the verbatim and escape sequence input routines use size_t's to hold lengths, and use a properly cast nrealloc() instead of an uncast realloc() in the former
git-svn-id: svn://svn.savannah.gnu.org/nano/trunk/nano@1717 35c25a1d-7b9e-4130-9fde-d3aeb78583b8

+++ src/nano.c
@@ -998,8 +998,8 @@ void do_char(char ch)
 int do_verbatim_input(void)
 {
     int *verbatim_kbinput;	/* Used to hold verbatim input. */
-    int verbatim_len;		/* Length of verbatim input. */
-    int i;
+    size_t verbatim_len;	/* Length of verbatim input. */
+    size_t i;
 
     statusbar(_("Verbatim input"));
     verbatim_kbinput = get_verbatim_kbinput(edit, &verbatim_len, 1);
